Rewrite view logic with ejs templates
The original client side view logic involves sending an HTML file
to the client with scripts calling back to the server via socket.io.
While this had the benefit of not requiring page refreshes for
updating data, the approach was clunky at best.

For now, the new view logic is written with ejs templates, which are
rendered and populated with match data on page load, and the server
delivers a complete, static page to the client. In the future, it may
be worth considering implementing auto-refresh functionality through
less clunky approaches than previously done.